requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
confusable-homoglyphs==3.3.1
croniter==3.0.3
cryptography==43.0.3
defusedxml==0.7.1
Django==5.2
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-recaptcha==4.0.0
django-registration==3.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
docopt==0.6.2
feedparser==6.0.11
html2text==2024.2.26
idna==3.10
lxml==5.3.0
Markdown==3.7
markdown-del-ins==1.0.0
mwparserfromhell==0.6.6
oauthlib==3.2.2
pipreqs==0.4.13
psycopg==3.2.3
pudb==2024.1.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.4.0
sgmllib3k==1.0.0
six==1.16.0
social-auth-core==4.5.4
sqlparse==0.5.1
urllib3==2.2.3
urwid==2.6.15
vk==3.0
yarg==0.1.9
ruff==0.12.0
mypy==1.16.1
